cmd/go: on windows, if args are larger than limit then use response file
This modifies golang's tools to check if the total number of
arguments will result in a commandline larger than 32k. If
we're on windows, this will then replace the commandline args
with a response file which should bypass the 32k limitation.

This was done by adding a function that copies the arguments
into a response file if the total length of the args is larger
than the windows limit. Each of the arguments written to the
response file are newline delimited, double-quoted, and escaped.

Closes issue #18468.

src/cmd/go/internal/work/exec.go: added buildLongCommand and escapeString