debug/pe: added windows-specific test against atmfd imports
This adds a test against atmfd as suggested by Alex Brainman
that proves that file.go has the ability to parse an import
table residing in a section not named .idata.

This test is windows-only and since the atmfd driver itself is
actually provided by Adobe but bundled with all versions of
windows, we're forced to search for the file in the current
PATH and then try to read the symbols from it. If the file
isn't found, the test fails.

src/debug/pe/file_test.go: added test